Aprendizaje/requirements.txt
azure-functions
pyodbc
pandas
sklearn